requirements.txt
Flask==2.0.1
Flask-WTF==1.2.1
Werkzeug==2.0.1
python-dotenv==1.0.1
Pillow==8.3.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1 